+++ crate/server/src/tests/ttlv_tests/get_attribute_1_4.rs
@@ -41,7 +41,7 @@ pub(crate) fn get_attributes(client: &SocketClient, key_id: &str) {
         request_header: RequestMessageHeader {
             protocol_version: ProtocolVersion {
                 protocol_version_major: 1,
-                protocol_version_minor: 1,
+                protocol_version_minor: 4,
             },
             batch_count: 1,
             ..Default::default()
@@ -72,7 +72,7 @@ pub(crate) fn get_attributes(client: &SocketClient, key_id: &str) {
         response.response_header.protocol_version,
         ProtocolVersion {
             protocol_version_major: 1,
-            protocol_version_minor: 1,
+            protocol_version_minor: 4,
         }
     );
     assert_eq!(response.batch_item.len(), 1);